requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
opencv-python-headless==4.10.0.84
numpy==2.2.1
Pillow==11.1.0
pytesseract==0.3.13
google-genai==1.7.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
python-dotenv==1.0.1
jinja2==3.1.5
aiofiles==24.1.0
